Article2Video/requirements.txt
fastapi
uvicorn[standard]
python-multipart
requests
beautifulsoup4
Pillow
numpy
openai
pydantic
python-dotenv 